pyproject.toml
[project]
name = "api-fun"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google>=3.0.0",
    "google-genai>=1.52.0",
    "python-dotenv>=1.2.1",
]
